requirements.txt
lanchain
langchain-community
langchain-huggingface
langchain_pinecone
python-dotenv
pypdf
sentence-transformers
pinecone
streamlit